Move bootstrap subcommand module
