+++ relayer/crates/starknet-cli/src/commands/mod.rs
@@ -1,4 +1,5 @@
 pub mod all;
+pub mod bootstrap;
 pub mod create;
 pub mod query;
 pub mod start;
