+++ relayer/crates/starknet-cli/src/impls/bootstrap/mod.rs
@@ -1,3 +1,2 @@
 pub mod osmosis_chain;
 pub mod starknet_chain;
-pub mod subcommand;
